requirements.txt
backcall==0.1.0
decorator==4.3.2
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pyserial==3.4
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
